Split/refactor database architecture
