+++ backend/HeatingDataMonitor.Database.Read/NpgsqlNotificationHeatingDataReceiver.cs
@@ -10,7 +10,7 @@ using Npgsql;
 
 // ReSharper disable ContextualLoggerProblem
 
-namespace HeatingDataMonitor.Database;
+namespace HeatingDataMonitor.Database.Read;
 
 // TODONE VERY IMPORTANT Test if this works for long durations. If not we may need keepalives.
 //      I tested this in isolation and didn't encounter any issue. We'll have to see how it behaves in the actual backend.
