+++ backend/TestingGrounds/TestingGrounds.csproj
@@ -8,7 +8,10 @@
     </PropertyGroup>
 
     <ItemGroup>
+      <ProjectReference Include="..\HeatingDataMonitor.Database.Read\HeatingDataMonitor.Database.Read.csproj" />
+      <ProjectReference Include="..\HeatingDataMonitor.Database.Write\HeatingDataMonitor.Database.Write.csproj" />
       <ProjectReference Include="..\HeatingDataMonitor.Database\HeatingDataMonitor.Database.csproj" />
+      <ProjectReference Include="..\HeatingDataMonitor.Receiver.Shared\HeatingDataMonitor.Receiver.Shared.csproj" />
     </ItemGroup>
 
     <ItemGroup>
