+++ backend/heating_data_monitor_architecture.drawio
@@ -1,0 +1,1 @@
+<mxfile host="app.diagrams.net" modified="2022-03-03T09:47:45.674Z" agent="5.0 (X11)" etag="V-v2tREyVKMTjbQCsqav" version="16.6.5" type="device"><diagram id="s8dsEkdGTyLwh6e7yYzE" name="Page-1">5Zpbj5s6EMc/TaT2oSvuYR+7l16k02rVVO1uX44c7Aa3BEfG2STn09cEO8AQchw2C6lWyoM92ED+/nnGYzNyr+fr9xwt4k8Mk2TkWHg9cm9GjmN7jjPKfxbeFJYgDArDjFOsGpWGCf2PKKOlrEuKSVZrKBhLBF3UjRFLUxKJmg1xzlb1Zj9ZUn/qAs1IwzCJUNK0fqdYxIU19K3S/oHQWayfbFvqyhzpxsqQxQizVcXk3o7ca86YKErz9TVJcvG0LkW/dy1Xdy/GSSpMOnxeTqa/f4Sf0P2vu/WYXH/8N6VvnLF6ObHR/5hgKYCqMi5iNmMpSm5L6xVnyxST/LaWrJVt/mFsIY22NP4iQmzUaKKlYNIUi3mirpI1Ffeqe15+yMsXvqrdrCuXbja6kgq+ua9WKr3yatltW9P9mjIp5TK25BE5oI3GDfEZEYfaKaJz4SpPUKPwnrA5kS8kG3CSIEEf62QhBehs1051fcs52lQaLBhNRVa5811ukA3UXAsVZ2qi2T7A4XBzP6w1l4Xi+bpW+SOlaUvYEbSpv/yIkqUSYeQEidT2CtNHWZzlxS8kIlIkri/JJ1Wu7unwasFZRLLsdVsPAHiJb87iKqaCTBZoi8JKOrE6qq34yDcUZH1wvPVVB4yMrq9Kh7KzxRVnEljtiFQG6vhxcEzG4QpFv0mKcxLuPrYOBprngqXTbFEfiYE1D6xz09xtaH6DBJqijBwQy+pFLDeoi7UTYTCxPBNA71gmZlwuDpx3X+mcZNugbew0yFoQLoOWfExCpxzxzc5/TDnsaOxRDAYMoyzezYRnGD178NHzW1G/+M6lNoMDDz3y8MA74ZDrMF0+ah1m97UOC/6uddgYetPLwwsx2L6XlVjQPkW/EIQHn6Ewfg8/Q8ftim2z3mxwzWAgcAfXLGzX7Bslq+Elg5gNL9nlnpUPDAwpfpvvbMhalKAso1Fdl7qIT/LxveXavqGPdw1dfGX8/D3Dp23HRYKG6/YgPg7AovjfqteBGKBjxM7deeBGhTCNG50sMbeeC7sStYcqaf+DXUnaQ3WlcXLsXEPsvLPCzga0eNAbGWPngRvZPWO3b0PoGbCzO2FnDY7d+Kyw80HW5IYdsfPH4EZuz9jt2/96GnYdk6IuG+HdsTNNpM4ryPo2iI2XXbGDiVnfQba5BXgq7I70XF28ZHfsTM9R/LPCDoRG1+pIHVzZ+T1Dt28r9QWEWMcQOtM9o56gAyHW75pQeCDE7up9YdfcA34R2JmG2PCssIOR0e8aYgMQq/2eQ6xz2aSsx88ZOuW6fW2x2KZsmmYdxtA9aUPMbm5V6y8ELiax1AE3j+1eoWkmOIoEZeko/+iHb18XJW8EnRNZxkigrUlwSuSdDx1Y93QkBaegntrDHUmdfk/or8iSwhNPk35ceOACfuBWjrELBx8GNTaXntuF70vOG8f3nxlGX7dz+cWf93uDn1k4LzSxHRt6Ci3HubgKmGR0dhUwyTiZq5DV8sPYonn5ebF7+wc=</diagram></mxfile>
